Create InternalTransferOptions component
