+++ client/dashboard/domains/domain-transfer/index.tsx
@@ -16,19 +16,17 @@ import { useLocale } from '../../app/locale';
 import { domainQuery } from '../../app/queries/domain';
 import { domainLockMutation, domainTransferCodeMutation } from '../../app/queries/domain-transfer';
 import { domainRoute } from '../../app/router/domains';
-import { ActionList } from '../../components/action-list';
 import InlineSupportLink from '../../components/inline-support-link';
 import Notice from '../../components/notice';
 import { PageHeader } from '../../components/page-header';
 import PageLayout from '../../components/page-layout';
-import RouterLinkButton from '../../components/router-link-button';
 import { SectionHeader } from '../../components/section-header';
 import { DomainSubtype } from '../../data/domains';
 import { formatDate } from '../../utils/datetime';
 import { getTopLevelOfTld } from '../../utils/domain';
+import InternalTransferOptions from './internal-transfer-options';
 import SelectIpsTag from './select-ips-tag';
 
-const RESTRICTED_TRANSFER_TLDS = [ 'uk', 'fr', 'ca', 'de', 'jp' ];
 const TRANSFER_LOCK_GRACE_PERIOD_MS = 60 * 24 * 60 * 60 * 1000; // 60 days
 
 export default function DomainTransfer() {
@@ -72,77 +70,6 @@ export default function DomainTransfer() {
 				) }
 			</Notice>
 		);
-	};
-
-	const renderInternalTransferOptions = () => {
-		const actions = [];
-		if ( ! domain.is_domain_only_site && domain.can_transfer_to_any_user ) {
-			const description =
-				domain.subtype.id === DomainSubtype.DOMAIN_CONNECTION
-					? __( 'Transfer this domain connection to any administrator on this site.' )
-					: __( 'Transfer this domain to any administrator on this site.' );
-			actions.push(
-				<ActionList.ActionItem
-					key="transfer-to-another-user"
-					title={ __( 'Transfer to another user' ) }
-					description={ description }
-					actions={
-						<RouterLinkButton
-							variant="secondary"
-							size="compact"
-							to={ `/domains/${ domain.domain }/transfer/other-user` }
-						>
-							{ __( 'Continue' ) }
-						</RouterLinkButton>
-					}
-				/>
-			);
-		} else if (
-			! RESTRICTED_TRANSFER_TLDS.includes( getTopLevelOfTld( domain.domain ) ) &&
-			domain.can_transfer_to_other_site
-		) {
-			actions.push(
-				<ActionList.ActionItem
-					key="transfer-to-any-user"
-					title={ __( 'To another WordPress.com user' ) }
-					description={ __( 'Transfer this domain to another WordPress.com user' ) }
-					actions={
-						<RouterLinkButton
-							variant="secondary"
-							size="compact"
-							to={ `/domains/${ domain.domain }/transfer/any-user` }
-						>
-							{ __( 'Continue' ) }
-						</RouterLinkButton>
-					}
-				/>
-			);
-		}
-		if ( domain.can_transfer_to_other_site ) {
-			const description =
-				domain.subtype.id === DomainSubtype.DOMAIN_CONNECTION
-					? __( 'Transfer this domain connection to any site you are an administrator on.' )
-					: __( 'Transfer this domain to any site you are an administrator on.' );
-			actions.push(
-				<ActionList.ActionItem
-					key="transfer-to-another-site"
-					title={ __( 'To another WordPress.com site' ) }
-					description={ description }
-					actions={
-						<RouterLinkButton
-							variant="secondary"
-							size="compact"
-							to={ `/domains/${ domain.domain }/transfer/other-site` }
-						>
-							{ __( 'Continue' ) }
-						</RouterLinkButton>
-					}
-				/>
-			);
-		}
-		if ( actions.length > 0 ) {
-			return <ActionList>{ actions }</ActionList>;
-		}
 	};
 
 	const renderTransferMessage = () => {
@@ -281,7 +208,7 @@ export default function DomainTransfer() {
 	return (
 		<PageLayout size="small" header={ <PageHeader title={ __( 'Transfer' ) } /> }>
 			{ renderTransferInfo() }
-			{ isDomainTransferable && <>{ renderInternalTransferOptions() }</> }
+			{ isDomainTransferable && <InternalTransferOptions domain={ domain } /> }
 			{ isDomainTransferable && domain.subtype.id !== DomainSubtype.DOMAIN_CONNECTION && (
 				<>{ renderExternalTransferOptions() }</>
 			) }
